isolate props propagation in a hook
